+++ BiomeConqueror/Benefits/MistlandsPatch.cs
@@ -17,25 +17,28 @@ namespace BiomeConqueror.Benefits
 
         static void Postfix(ref Demister __instance)
         {
-            if (!ConfigurationFile.modEnabled.Value) return;
-            
+            if (!ConfigurationFile.modEnabled.Value || Player.m_localPlayer == null) return;
+
             try
             {
                 if (BiomeConquerorUtils.isQueenDefeatedForPlayer())
                 {
-                    var itemData = Player.m_localPlayer.GetInventory().GetEquippedItems().FirstOrDefault(i => i.m_dropPrefab.name == "Demister");
+                    var itemData = Player.m_localPlayer.GetInventory().GetEquippedItems()
+                        .FirstOrDefault(i => i.m_dropPrefab.name == "Demister");
 
                     if (!__instance.isActiveAndEnabled || itemData == null) return;
                     __instance.m_forceField.endRange = ConfigurationFile.queenBenefitEligibleRange.Value;
                 }
+
                 demisterRange = __instance.m_forceField.endRange;
 
                 //Update wisp light buff text
-                var demisterSE = Player.m_localPlayer.GetSEMan().GetStatusEffects().First(effect => effect.name == "Demister");
+                var demisterSE = Player.m_localPlayer.GetSEMan().GetStatusEffects()
+                    .First(effect => effect.name == "Demister");
                 demisterSE.m_name = $"$item_demister" + ": {demisterRange}m.";
                 Logger.Log($"demister buff text updated to {demisterSE.m_name}");
             }
-            catch (Exception ex) {}
+            catch (Exception ex){}
         }
     }